+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Rename `:default` analyzer to `:content_type` in `determine_mime_type` plugin (@janko-m)
+
 * Don't display a warning when `determine_mime_type` plugin is loaded with `:default` analyzer (@janko-m)
 
 * Exclude media type parameters when copying `IO#content_type` into `mime_type` metadata (@janko-m)